src/requirements.txt
pywinauto>=0.6.8
pyautogui>=0.9.54
pyperclip>=1.8.2
pillow>=9.5.0
pyyaml>=6.0.1
watchdog>=3.0.0
python-dotenv==1.0.0 